+++ src/components/ControlPanel/ControlPanel.scss
