+++ src/components/FilterDualSlider/FilterDualSlider.scss
