+++ src/components/FilterDualSlider/FilterDualSlider.tsx
@@ -1,7 +1,7 @@
 import { FormEvent, useEffect, useState } from "react";
 import { useSearchParams } from "react-router-dom";
 import { IProduct } from "../../interfaces/products";
-import './dualSlider.scss';
+import './FilterDualSlider.scss';
 
 interface IFilterDualSliderProps {
   filterName: 'price' | 'stock';
